Implement JSON pseudo-compression
When using GET requests, the safe size limit for requests is 2K. Basic JSON encoding has a high overhead: the key names are verbose and repeated, and the frequent " characters are URL-encoded as %22.

Apply a basic compression scheme, by specifying a hard-coded dictionary of key shorthands and swapping " and / before sending (for avoid the URL-encoding overhead).